final/back-server/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.14.0
django-filter==23.2
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.9.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
toposort==1.10
urllib3==2.0.2
